Update to hapi lab and code dependencies
We think the issue is that hapi was listed x2 as a dev dependency. Once, as the new @hapi/hapi but the second time under the old `hapi` namespace. Along with that old `lab` and `Code` were listed as the test dependencies.

The old Hapi looks like it was bringing in a version of catbox-memory that depended on big-time but that wasn't actually getting installed. We've removed that second entry but to be on the safe side we have also updated the versions of Lab and Code to use those under the `@hapi` namespace.