Make package.json consistent
https://github.com/DEFRA/water-abstraction-team/issues/31

With many repos comes inconsistent package.json files it seems. We also have some minor errors, most notably on the code used for the license.

Finally, we have a lot of scripts in some of the packages which none of the current dev team uses.

So, we just need to go through the repos and ensure they are consistent, correct and some of the script noise is removed.